backend/requirements.txt
blinker==1.9.0
click==8.3.2
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg[binary]==3.3.3
PyJWT==2.12.1
python-dotenv==1.2.2
Werkzeug==3.1.8